+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Tất cả những thay đổi đáng chú ý của dự án kiencang/SI-Prompt-
 Định dạng dựa trên [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 và dự án này tuân thủ [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.46] - 2026-05-31
+### Fixed
+- Điều chỉnh vị trí button mục lục & chỉnh giao diện ở gần top hơn.
+- Điều chỉnh CSS cho các toggle khi được active.
+- Thêm Thinking level HIGH vào quá trình dịch.
+
 ## [1.0.45] - 2026-05-28
 ### Fixed
 - Chuẩn hóa style CSS để quá trình biên dịch thành app giảm thiểu rủi ro lỗi.